+++ cmd/cli/plugin/package/package_installed_delete.go
@@ -58,7 +58,7 @@ func packageUninstall(cmd *cobra.Command, args []string) error {
 	go pkgClient.UninstallPackage(packageInstalledOp, pp)
 
 	initialMsg := fmt.Sprintf("Uninstalling package '%s' from namespace '%s'", packageInstalledOp.PkgInstallName, packageInstalledOp.Namespace)
-	if err := displayProgress(initialMsg, pp); err != nil {
+	if err := DisplayProgress(initialMsg, pp); err != nil {
 		if err.Error() == tkgpackagedatamodel.ErrPackageNotInstalled {
 			log.Warningf("\npackage '%s' is not installed in namespace '%s'. Cleaned up related resources", packageInstalledOp.PkgInstallName, packageInstalledOp.Namespace)
 			return nil
