+++ cmd/cli/plugin/package/package_installed_update.go
@@ -54,7 +54,7 @@ func packageUpdate(cmd *cobra.Command, args []string) error {
 	go pkgClient.UpdatePackage(packageInstalledOp, pp)
 
 	initialMsg := fmt.Sprintf("Updating package '%s'", packageInstalledOp.PkgInstallName)
-	if err := displayProgress(initialMsg, pp); err != nil {
+	if err := DisplayProgress(initialMsg, pp); err != nil {
 		if err.Error() == tkgpackagedatamodel.ErrPackageNotInstalled {
 			log.Warningf("\npackage '%s' is not among the list of installed packages in namespace '%s'. Consider using the --install flag to install the package", packageInstalledOp.PkgInstallName, packageInstalledOp.Namespace)
 			return nil
